pyproject.toml
[project]
name = "ml-article-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "transformers>=4.51.0",
    "peft>=0.18.0",
    "torch>=2.0.0",
    "accelerate>=0.33.0",
    "bitsandbytes>=0.46.1",
    "flask>=3.0.0",
    "python-docx>=1.1.0",
]

[project.scripts]
ml-article-writer = "article_writer.writer.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["article_writer*"]
